requirements.txt
Django==3.1.3
mysqlclient==2.0.1
djangorestframework==3.12.1